% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covDcor}
\alias{covDcor}
\title{Returns the theoretical covariance of empirical correlations.}
\usage{
covDcor(r)
}
\arguments{
\item{r}{a correlation matrix}
}
\value{
Returns the theoretical covariance of empirical correlations.
}
\description{
Returns the theoretical covariance of empirical correlations.
}
\examples{
p <- 10
corr_theo <- diag(1,p)
corr_theo[2:p,] <- 0.3
corr_theo[,2:p] <- 0.3
covDcor(corr_theo)
}
\references{
Aitkin, M. A. (1969). Some tests for correlation matrices. Biometrika, 443-446.
}
\seealso{
covDcorNorm
}
