% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unvectorize}
\alias{unvectorize}
\title{Returns an upper-triangle matrix, without the diagonal, containing the elements of a given vector.}
\usage{
unvectorize(vect)
}
\arguments{
\item{vect}{A vector containing the upper triangle of a matrix, without the diagonal}
}
\value{
Returns an upper-triangle matrix where each entry is given by the vector containing the upper triangle of a matrix, without the diagonal.
}
\description{
Returns an upper-triangle matrix, without the diagonal, containing the elements of a given vector.
}
\examples{
unvectorize(1:10)
}
\seealso{
vectorize
}
