% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorize}
\alias{vectorize}
\title{Returns a vector containing the upper triangle of a matrix, without the diagonal.}
\usage{
vectorize(mat)
}
\arguments{
\item{mat}{a square matrix}
}
\value{
Returns a vector containing the upper triangle of a matrix, without the diagonal.
}
\description{
Returns a vector containing the upper triangle of a matrix, without the diagonal.
}
\examples{
vectorize(matrix(1:9,3,3))
}
\seealso{
unvectorize
}
